final refactoring after review (bigInteger on migration)

+++ database/migrations/2022_02_28_115109_create_study_classes_lectures_table.php
@@ -15,9 +15,9 @@ return new class extends Migration
     {
         Schema::create('study_classes_lectures', function (Blueprint $table) {
             $table->index(['study_class_id', 'lecture_id']);
-            $table->integer('study_class_id')->unsigned();
+            $table->bigInteger('study_class_id')->unsigned();
             $table->foreign('study_class_id')->references('id')->on('study_classes');
-            $table->integer('lecture_id')->unsigned();
+            $table->bigInteger('lecture_id')->unsigned();
             $table->foreign('lecture_id')->references('id')->on('lectures');
             $table->integer('sequence')->unsigned();
         });
